+++ game-deadasdisco/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - None
 
+## [0.1.1] - 2026-04-13
+
+- Fixed: error no longer shown when launching demo version
+
 ## [0.1.0] - 2026-01-13
 
 - Initial release