+++ game-deathstranding2onthebeach/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - None Planned
 
+## [0.2.0] - 2026-04-13
+
+- Added: HFW Mod Manager integration for installing .core and .stream mods
+
 ## [0.1.0] - 2026-04-08
 
 - Initial Release