+++ game-horizonzerodawnremastered/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.2.0] - 2026-04-13
+
+- Added: HFW Mod Manager integration for installing .core and .stream mods
+
 ## [0.1.2]
 
 - Initial release