+++ game-styxbladesofgreed/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - None
 
+## [0.1.1] - 2026-04-13
+
+- Fixed: Corrected Config folder name to "WindowsClient"
+
 ## [0.1.0] - 2026-02-19
 
 - Initial release